// src/types/User.ts
export interface User {
  id: string;
  name: string;
  email: string; 
  password: string;
  provider: string;
  roleCode: string;
  businessUnitId: string;
  username: string;
  status: string;
  firstName: string;
  lastName: string;
  department: string;
  phoneNumber: string;
  activeDirectoryId: string;
  createdAt: string;
  modifiedAt: string;
  createdById: string;
  modifiedById: string;
}